move CLI handling to cli.py
