+++ tests/test_gates.py
@@ -310,3 +310,21 @@ def test_phase_shift():
     h.compare_matrices(ps2, expected2)
     h.compare_matrices(ps3, expected3)
     h.compare_matrices(ps4, expected4)
+
+# vec=DefaultMatrix([[0] for _ in range(4)])
+# vec[0] = [1]
+# print(QFT_Gate(2)*vec)
+# phase = 0.25
+# unitary = DefaultMatrix([[1,0],[0,cmath.exp(2j*cmath.pi*phase)]])
+# eigenvec = DefaultMatrix([[0],[1]])
+# PE = PhaseEstimation(2,unitary,eigenvec)
+# PE.run()
+# print(PE.measure())
+
+# vec=DefaultMatrix([[0] for _ in range(8)])
+# vec[6] = [1]
+# print( Inverse_QFT_Gate(3)*vec)
+
+# print(g.control_phase(3,[0],1,math.pi))
+# print('\n')
+# print(g.control_phase(3,[2],1,math.pi))
